quick_tube_backend/requirements.txt
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
transformers
torch